last of refactoring for now
