+++ shared/src/persistence/dynamo/cases/getCaseByDocketNumber.js
@@ -1,4 +1,6 @@
-const { getRecordViaMapping } = require('../../awsDynamoPersistence');
+const {
+  getRecordViaMapping,
+} = require('../../dynamo/helpers/getRecordViaMapping');
 const { stripWorkItems } = require('../../dynamo/helpers/stripWorkItems');
 const { stripInternalKeys } = require('../../dynamo/helpers/stripInternalKeys');
