+++ shared/src/persistence/dynamo/cases/getCasesByDocumentId.js
@@ -1,5 +1,6 @@
-const { getRecordsViaMapping } = require('../../awsDynamoPersistence');
-
+const {
+  getRecordsViaMapping,
+} = require('../../dynamo/helpers/getRecordsViaMapping');
 const { stripInternalKeys } = require('../../dynamo/helpers/stripInternalKeys');
 const { stripWorkItems } = require('../../dynamo/helpers/stripWorkItems');
