+++ shared/src/persistence/dynamo/documents/syncDocuments.test.js
@@ -7,7 +7,6 @@ const applicationContext = {
   environment: {
     stage: 'local',
   },
-  getDynamoClient: () => client,
   filterCaseMetadata: ({ cases }) => cases,
   isAuthorizedForWorkItems: () => true,
 };
