+++ shared/src/persistence/dynamo/helpers/getRecordsViaMapping.js
@@ -1,0 +1,36 @@
+const { stripInternalKeys } = require('./stripInternalKeys');
+const client = require('../../dynamodbClientService');
+
+exports.getRecordsViaMapping = async ({
+  applicationContext,
+  key,
+  type,
+  isVersioned = false,
+}) => {
+  const TABLE = `efcms-${applicationContext.environment.stage}`;
+
+  const mapping = await client.query({
+    applicationContext,
+    TableName: TABLE,
+    ExpressionAttributeNames: {
+      '#pk': 'pk',
+    },
+    ExpressionAttributeValues: {
+      ':pk': `${key}|${type}`,
+    },
+    KeyConditionExpression: '#pk = :pk',
+  });
+
+  const ids = mapping.map(metadata => metadata.sk);
+
+  const results = await client.batchGet({
+    applicationContext,
+    tableName: TABLE,
+    keys: ids.map(id => ({
+      pk: id,
+      sk: isVersioned ? '0' : id,
+    })),
+  });
+
+  return stripInternalKeys(results);
+};
