+++ shared/src/persistence/dynamo/workitems/getSentWorkItemsForUser.js
@@ -1,4 +1,6 @@
-const { getSortRecordsViaMapping } = require('../../awsDynamoPersistence');
+const {
+  getSortRecordsViaMapping,
+} = require('../../dynamo/helpers/getSortRecordsViaMapping');
 const moment = require('moment');
 
 exports.getSentWorkItemsForUser = ({ userId, applicationContext }) => {
